Rename track_request/TrackRequest to TrackRequest/index
