+++ frontend/src/features/map/overlays/map_menu/TrackRangeModal.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
 import { COLORS } from '../../../../constants/constants'
 import { Modal } from 'rsuite'
-import DateRange from '../../../vessel_sidebar/actions/track_request/DateRange'
+import DateRange from '../../../vessel_sidebar/actions/TrackRequest/DateRange'
 import StyledModalHeader from '../../../commonComponents/StyledModalHeader'
 
 /**
