+++ frontend/src/features/vessel_sidebar/VesselSidebar.tsx
@@ -6,8 +6,8 @@ import { MapComponentStyle } from '../commonStyles/MapComponent.style'
 import AnimateToTrack from './actions/animate_to_track/AnimateToTrack'
 import HideNonSelectedVessels from './actions/hide_non_selected_vessels/HideNonSelectedVessels'
 import ShowFishingActivitiesOnMap from './actions/show_fishing_activities/ShowFishingActivitiesOnMap'
-import { ExportTrack } from './actions/track_request/ExportTrack'
-import { TrackRequest } from './actions/track_request/TrackRequest'
+import { TrackRequest } from './actions/TrackRequest'
+import { ExportTrack } from './actions/TrackRequest/ExportTrack'
 import { VesselSidebarBody } from './VesselSidebarBody'
 import VesselSidebarTabs from './VesselSidebarTabs'
